+++ src/components/index.ts
@@ -3,4 +3,5 @@ export { default as Container } from './Container'
 export { default as ControlledSelect } from './ControlledSelect/ControlledSelect'
 export { default as ControlledTextInput } from './ControlledTextInput'
 export { default as Icon } from './Icon'
+export { default as InfographicItem } from './InfographicItem'
 export { default as Link } from './Link'
